Item presentation (#40)
* Change items source to local json items.json, AB#759

* Fix up/down pick bug, AB#759

* fix item disappearing issue, AB#759
