+++ src/main/resources/objects/potion.png
